# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Users API",
        "description": "API to manage students",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Retrieve a list of all users",
                "operationId": "b520321f3c073c07331f6bd4e979bd25",
                "responses": {
                    "200": {
                        "description": "List of users"
                    },
                    "404": {
                        "description": "No user found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Create a new student with provided name and age",
                "operationId": "bdf6245eeea671d2a0cb9abde9183227",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firs_name",
                                    "firs_name",
                                    "email",
                                    "number",
                                    "password"
                                ],
                                "properties": {
                                    "firs_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Edit a user",
                "description": "Update the details of a user",
                "operationId": "f6171afbcd09eef03dd7bc52fefa7333",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the users to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "description": "Update the details of a user",
                "operationId": "db41c9978a4259fa8e3abe0d652d9fa2",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID of the users to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firs_name",
                                    "firs_name",
                                    "email",
                                    "number",
                                    "password"
                                ],
                                "properties": {
                                    "firs_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "description": "Delete a users by its ID",
                "operationId": "50197234e2a2b0020d3422b73b2c6ce2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the student to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "user deleted"
                    },
                    "404": {
                        "description": "user not found"
                    }
                }
            }
        }
    }
}